Change rating from double to int

+++ backend/src/main/java/com/rateNUS/backend/hostel/HostelService.java
@@ -44,7 +44,7 @@ public class HostelService {
     }
 
     @Transactional
-    public void updateHostel(long hostelId, double rating, boolean hasNewComment) {
+    public void updateHostel(long hostelId, int rating, boolean hasNewComment) {
         if (!hasNewComment) {
             return;
         }
